Remove declaration of direct Wayland support
Similar what was experienced in #11, the main game session fails to start
on Wayland (main menu works). One workaround is to select `osggraph` as
the graphics engine, but I reckon there's a reason the other `ssggraph`
is selected as default?

Anyway, to make the game start correctly and with ssggraph out of the box,
_direct_ wayland support has to be disabled until the underlying issue
w/ `ssggraph` and Wayland is fixed.